+++ src/main/scala/za/co/absa/spark_metadata_tool/model/AppConfig.scala
@@ -33,4 +33,4 @@ sealed trait Mode
 case object FixPaths                                                              extends Mode
 case object Merge                                                                 extends Mode
 case object CompareMetadataWithData                                               extends Mode
-final case class CreateMetadata(maxMicroBatchNumber: Long, compactionNumber: Int) extends Mode
+final case class CreateMetadata(maxMicroBatchNumber: Int, compactionNumber: Int) extends Mode
